duha/requirements.txt
openai == 1.3.7
PyPDF2 == 3.0.1
python-dotenv == 1.0.0
#langchain == 0.0.149
chromadb == 0.4.18
tiktoken == 0.5.2
djangorestframework == 3.14.0
python-dotenv == 1.0.0
tqdm == 4.66.1
sentence-transformers == 2.2.2
streamlit == 1.29.0
django-cors-headers == 4.3.1
pypdf == 3.17.4
langchain == 0.0.345